server/requirements.txt
fastapi~=0.121.2
pydantic~=2.12.4
uvicorn~=0.38.0
sqlmodel~=0.0.27
pyjwt~=2.10.1
passlib[bcrypt]
python-multipart
python-dotenv
typing_extensions~=4.15.0
SQLAlchemy~=2.0.44